[Test] Move tests
